python/uv/pyproject.toml
[project]
name = "uv-vulnerable-demo"
version = "0.1.0"
description = "Demo of vulnerable Python dependencies using uv"
requires-python = ">=3.10"
dependencies = [
    "flask==1.0.2",
    "pyyaml==5.3.1",
    "requests==2.19.1",
    "sqlalchemy==1.3.0",
    "werkzeug==0.14.1",
    "jinja2==2.10.1",
    "cryptography==43.0.0",
]
